+++ CHANGELOG.md
@@ -66,6 +66,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Variant loading filter based on rank score and category (#98)
 - Parse custom images for STR variants (#102)
 - Support loading VCFs by category and research status (#111)
+- Tests for clinical and research VCF category selection (#113)
 ### Changed
 - Structure of parsed entities, to reflect documents saved in the database (#20)
 - Improved and simplified code (#22)